+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Add `LiteSVM::with_precompiles`.
+
 ## [0.3.0] - 2024-10-12
 
 ### Added