+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+## [0.4.1] - 2025-01-23
+
+- Fix Solana dependencies ([#119](https://github.com/LiteSVM/litesvm/pull/119)).
+
 ## [0.4.0] - 2024-12-30
 
 ### Changed
@@ -61,7 +65,8 @@
 
 - Initial release.
 
-[Unreleased]: https://github.com/LiteSVM/litesvm/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/LiteSVM/litesvm/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/LiteSVM/litesvm/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/LiteSVM/litesvm/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/LiteSVM/litesvm/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/LiteSVM/litesvm/compare/v0.2.0...v0.2.1